# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Events System API",
        "contact": {
            "email": "mahmoudnada5050@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/events/{post}/comments": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "Add new Comment To Post",
                "parameters": [
                    {
                        "name": "type",
                        "in": "header",
                        "description": "1 => attendee , 2=> speaker",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "desc",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/events/{post}/comments/{comment}": {
            "put": {
                "tags": [
                    "comments"
                ],
                "summary": "update Comment",
                "parameters": [
                    {
                        "name": "type",
                        "in": "header",
                        "description": "1 => attendee , 2=> speaker",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "desc",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "_method:put",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
        "/events/{event}/posts": {
            "get": {
                "tags": [
                    "events"
                ],
                "summary": "Get single speaker",
                "parameters": [
                    {
                        "name": "type",
                        "in": "header",
                        "description": "1 => attendee , 2=> speaker",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "Add new Post To Event",
                "parameters": [
                    {
                        "name": "type",
                        "in": "header",
                        "description": "1 => attendee , 2=> speaker",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "desc",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "photo",
                        "in": "formData",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/events/{event}/posts/{post}/update": {
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "update Event Post",
                "parameters": [
                    {
                        "name": "type",
                        "in": "header",
                        "description": "1 => attendee , 2=> speaker",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "desc",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "photo",
                        "in": "formData",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}